// nextjs-app/sanity.types.ts
export type Service = {
  _id: string;
  title: string;
  description: string;
  image?: {
    asset?: {
      _ref: string;
      url: string;
    };
  };
};